+++ UriTemplateTest.php
@@ -35,7 +35,7 @@ final class UriTemplateTest extends TestCase
 
         $uriTemplate = new UriTemplate($template, $variables);
 
-        self::assertSame($template, (string) $uriTemplate->template);
+        self::assertSame($template, (string) $uriTemplate);
     }
 
     public function testGetDefaultVariables(): void
@@ -59,12 +59,12 @@ final class UriTemplateTest extends TestCase
         ];
 
         $uriTemplate = new UriTemplate($template, $variables);
-        self::assertSame($expectedVariables, [...$uriTemplate->defaultVariables]);
-        self::assertFalse($uriTemplate->defaultVariables->isEmpty());
+        self::assertEquals($expectedVariables, [...$uriTemplate->getDefaultVariables()]);
+        self::assertCount(5, $uriTemplate->getDefaultVariables());
 
         $uriTemplateEmpty = new UriTemplate($template, []);
-        self::assertSame([], [...$uriTemplateEmpty->defaultVariables]);
-        self::assertTrue($uriTemplateEmpty->defaultVariables->isEmpty());
+        self::assertSame([], [...$uriTemplateEmpty->getDefaultVariables()]);
+        self::assertCount(0, $uriTemplateEmpty->getDefaultVariables());
     }
 
     public function testWithDefaultVariables(): void
@@ -79,9 +79,9 @@ final class UriTemplateTest extends TestCase
         $altTemplate = $uriTemplate->withDefaultVariables($variables);
         $newAltTemplate = $uriTemplate->withDefaultVariables($newAltVariables);
 
-        self::assertEquals($altTemplate->defaultVariables, $uriTemplate->defaultVariables);
-        self::assertEquals($newAltTemplate->defaultVariables, $uriTemplate->defaultVariables);
-        self::assertNotEquals($newTemplate->defaultVariables, $uriTemplate->defaultVariables);
+        self::assertEquals($altTemplate->getDefaultVariables(), $uriTemplate->getDefaultVariables());
+        self::assertEquals($newAltTemplate->getDefaultVariables(), $uriTemplate->getDefaultVariables());
+        self::assertNotEquals($newTemplate->getDefaultVariables(), $uriTemplate->getDefaultVariables());
     }
 
     /**
@@ -89,7 +89,7 @@ final class UriTemplateTest extends TestCase
      */
     public function testGetVariableNames(string $template, array $expected): void
     {
-        self::assertSame($expected, (new UriTemplate($template))->template->variableNames);
+        self::assertSame($expected, (new UriTemplate($template))->getVariableNames());
     }
 
     public static function expectedVariableNames(): iterable
@@ -119,7 +119,7 @@ final class UriTemplateTest extends TestCase
      */
     public function testExpandsUriTemplates(string $template, string $expectedUriString, array $variables): void
     {
-        self::assertSame($expectedUriString, (new UriTemplate($template))->expand($variables)->toString());
+        self::assertSame($expectedUriString, (new UriTemplate($template))->expand($variables));
     }
 
     public static function templateExpansionProvider(): iterable
@@ -255,11 +255,11 @@ final class UriTemplateTest extends TestCase
             'more'     => ['fun', 'ice cream'],
             'foo[]' => ['fizz', 'buzz'],
         ];
-        $expectedUri = 'http://example.com/foo/bar/one,two?query=test&more=fun&more=ice%20cream&foo%5B%5D=fizz&foo%5B%5D=buzz';
+        $expectedUri = 'http://example.com/foo/bar/one,two?query=test&more=fun&more=ice%20cream&foo[]=fizz&foo[]=buzz';
 
         $uriTemplate = new UriTemplate($template);
-        self::assertSame($expectedUri, $uriTemplate->expand($variables)->toString());
-        self::assertSame($expectedUri, $uriTemplate->expandOrFail($variables)->toString());
+        self::assertSame($expectedUri, $uriTemplate->expand($variables));
+        self::assertSame($expectedUri, $uriTemplate->expandOrFail($variables));
     }
 
     public function testDisallowNestedArrayExpansion(): void
@@ -296,11 +296,11 @@ final class UriTemplateTest extends TestCase
             'more' => ['fun', 'ice cream'],
             'foo[]' => ['fizz', 'buzz'],
         ];
-        $expectedUri = 'http://example.com/foo/bar/one,two?query=test&more=fun&more=ice%20cream&foo%5B%5D=fizz&foo%5B%5D=buzz';
+        $expectedUri = 'http://example.com/foo/bar/one,two?query=test&more=fun&more=ice%20cream&foo[]=fizz&foo[]=buzz';
 
         $uriTemplate = new UriTemplate($template, $defaultVariables);
-        self::assertSame($expectedUri, $uriTemplate->expand($variables)->toString());
-        self::assertSame($expectedUri, $uriTemplate->expandOrFail($variables)->toString());
+        self::assertSame($expectedUri, $uriTemplate->expand($variables));
+        self::assertSame($expectedUri, $uriTemplate->expandOrFail($variables));
     }
 
     public function testExpandWithDefaultVariablesWithOverride(): void
@@ -320,8 +320,8 @@ final class UriTemplateTest extends TestCase
         ];
 
         self::assertSame(
-            'http://example.com/bar/baz/one,two?query=test&more=fun&more=ice%20cream&foo%5B%5D=fizz&foo%5B%5D=buzz',
-            (new UriTemplate($template, $defaultVariables))->expand($variables)->toString()
+            'http://example.com/bar/baz/one,two?query=test&more=fun&more=ice%20cream&foo[]=fizz&foo[]=buzz',
+            (new UriTemplate($template, $defaultVariables))->expand($variables)
         );
     }
 
@@ -351,7 +351,7 @@ final class UriTemplateTest extends TestCase
         $template = '{foo}/{foo}';
         $data = ['foo' => 'foo'];
 
-        self::assertSame('foo/foo', (new UriTemplate($template, $data))->expand()->toString());
+        self::assertSame('foo/foo', (new UriTemplate($template, $data))->expand());
     }
 
     public function testExpandOrFailIfVariablesAreMissing(): void